Support for making multiple overlays in the same scratch path
* The work and upper dirs were hardcoded previously so if you had tried to
create more than one overlay in a mountpoint (/dev/sda / container scratch volume)
it would fail. This change just adds a simple check to see if workdir and upperdir
already exist and will create a unique name for them with a generated ID. This is the
same method cri uses to generate container IDs.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>