Add cifs mount and loopback device modify calls
To facilitate teleport enabled images from being used for LCOW, add two new modify____
branches for ModifyHostSettings.

* Add cifs mounting logic and associated calls.

* Add loopback device mounting logic and associated calls.

* Add `DoInNetNS` to be able to execute a function inside a specific network namespace. This is
needed as the common LCOW scenario has the only available network adapter not exposed to the
default network namespace. Therefore, to carry out a cifs mount for example, we need to join
this network namespace, perform the action, and then leave on the way out.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>